restructure pkg
