+++ integrations/agent_pack/src/haystack_integrations/agent_pack/deep_research/prompts.py
